% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cpselec}
\alias{cpselec}
\title{Select a CP values}
\usage{
cpselec(n = 10, len = 100, m = 2, ...)
}
\arguments{
\item{n}{number of cp intervals}

\item{len}{number of samples}

\item{m}{number of cp sample matrix rows (less than )}

\item{...}{}
}
\value{
Selected CP values
}
\description{
Select a CP values
}

