% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cregt}
\alias{cregt}
\title{Create a Generalization Table}
\usage{
cregt(dat, ori.dat, QI, TA, ...)
}
\arguments{
\item{dat}{\code{qigrp} object}

\item{ori.dat}{Original Data}

\item{QI}{Quasi-Identifiers}

\item{TA}{Target Attribute}

\item{...}{}
}
\value{
Generalization Table
}
\description{
Create a Generalization Table
}

