% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eclass}
\alias{eclass}
\title{Equivalence Class}
\usage{
eclass(dat, QI, TA, ...)
}
\arguments{
\item{dat}{\code{qigrp} object}

\item{QI}{Quasi-Identifiers}

\item{TA}{Target Attribute}

\item{...}{}
}
\value{
Equivalence Class and Size
}
\description{
Equivalence Class
}

