% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lanode}
\alias{lanode}
\title{Create a Lattice Node}
\usage{
lanode(obj, mc = FALSE, ...)
}
\arguments{
\item{obj}{\code{rpaclass} object}

\item{mc}{if TRUE \code{mclapply} applied, FALse \code{lapply} applied (default).}

\item{...}{}
}
\value{
Created Lattice Node using result of QI Classing
}
\description{
Create a Lattice Node
}

