% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parda}
\alias{parda}
\title{Partition Data}
\usage{
parda(dat, QI, TA, ...)
}
\arguments{
\item{dat}{Original data}

\item{QI}{Quasi-Identifiers}

\item{TA}{Target Attribute}

\item{...}{}
}
\value{
Partitioning Data by Quasi-Identifiers and Target Attribute
}
\description{
Partition Data
}
\examples{
data(AdultUCI, package='arules')
# TA (Target)
TA <- c('income')
# QI (Quasi-Identifiers)
QI <- c('age', 'workclass', 'education', 'marital.status', 'occupation', 'race', 'sex', 'native.country')
#QI <- c('age', 'workclass', 'education', 'marital.status')
# SA (Sensitive Attributes)
SA <- c('capital.gain', 'capital.loss')
# IS (Insensitive)
IS <- c('relationship', 'hours.per.week', 'education.num')
# Remove
RM <- c('fnlwgt')

parda(AdultUCI, QI, TA)
}

