% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qigrp}
\alias{qigrp}
\title{Quasi-Identifiers Grouping}
\usage{
qigrp(dat, obj, mcs = 4, ...)
}
\arguments{
\item{dat}{\code{parda} object}

\item{obj}{\code{lanode} object}

\item{...}{}
}
\value{
Grouped Data by each lattice node
}
\description{
Quasi-Identifiers Grouping
}

