% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resume}
\alias{resume}
\title{Suppression and Information Loss Metrics}
\usage{
resume(dat, kanon = 3, maxsup = 0.01, ...)
}
\arguments{
\item{dat}{\code{eclass} object}

\item{kanon}{k of k-Anonymity}

\item{maxsup}{threshold of k}

\item{...}{}
}
\value{
Suppression and Information Loss Metrics. Information Loss Metrics are Discernbility Metric (DM) and Entropy
}
\description{
Suppression and Information Loss Metrics
}

