% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rpaclass}
\alias{rpaclass}
\title{Classing QI by \code{rpart} function}
\usage{
rpaclass(dat, cp = 0.01, loss.mat = matrix(c(0, 1, 2, 0), ncol = 2),
  cval = 10, mc = FALSE, mcs = NULL, ...)
}
\arguments{
\item{dat}{\code{parda} object}

\item{cp}{Cost Complexity Parameter. \code{cpselec} object}

\item{loss.mat}{Loss matrix}

\item{cval}{number of cross-validations}

\item{mc}{if TRUE \code{mclapply} applied, FALse \code{lapply} applied (default).}

\item{mcs}{number of cores}

\item{...}{}
}
\value{
Classed a Quasi-Identifiers by CP using Entropy
}
\description{
Classing QI by \code{rpart} function
}

